Implement prometheus exporter with [Count, Distribution] Aggregation
Follow up of #181
Updates #168

Implements the Prometheus endpoint and now we can examine
data of the OpenCensus View Aggregation types of:
* CountAggregation
* DistributionAggregation

Also added a ViewRegistrar interface to stats so that
we can avoid the client having manually register views
to the exporter. With ViewRegistrar's RegisterView,
when a new View's Subscribe is invoked, the underlying
worker publishes the view to all Exporters that have
the RegisterView method. This way all the client has
to do is:
1. Create the Exporter.
2. Make measures and views
3. Invoke view.Subscribe() per view
thus simplifying our API.

And the sample output data after running the example
in examples/stats/prometheus/main.go

* Exhibit:

shows video count + processed byte data continuously
being generated and due to the nice API @rakyll
thought of, we can view the video data by
visiting the set endpoint, in a sample case /metrics

```shell
curl http://localhost:9999/metrics
```

```shell
 # HELP opencensus_video_count number of videos processed processed over
 # time
 # TYPE opencensus_video_count counter
 opencensus_video_count 1879
 # HELP opencensus_video_cum processed video size over time
 # TYPE opencensus_video_cum histogram
 opencensus_video_cum_bucket{le="0.005"} 372
 opencensus_video_cum_bucket{le="0.01"} 372
 opencensus_video_cum_bucket{le="0.025"} 372
 opencensus_video_cum_bucket{le="0.05"} 372
 opencensus_video_cum_bucket{le="0.1"} 372
 opencensus_video_cum_bucket{le="0.25"} 372
 opencensus_video_cum_bucket{le="0.5"} 372
 opencensus_video_cum_bucket{le="1"} 372
 opencensus_video_cum_bucket{le="2.5"} 372
 opencensus_video_cum_bucket{le="5"} 372
 opencensus_video_cum_bucket{le="10"} 372
 opencensus_video_cum_bucket{le="+Inf"} 1116
 opencensus_video_cum_sum 1.597752213504e+12
 opencensus_video_cum_count 1116
```