Fix accidental View name typo and add tests for plugin/grpc/grpcstats
Reported by @jcd2, I accidentally typo'd
  " rpc.io/client/error_count/cumulative"
instead of
  "grpc.io/client/error_count/cumulative"
in PR #217.

Added a test to ensure that we never regress/repeat this,
because the code has many moving parts and such a typo
or copy and paste errors would have lurked if not for
a sharp eye or an explicit error.