Implement a SanitizeNoClash that errors on same result but different inputs
Updates #173

A sanitizer that returns an error if the input produces a sanitization
that clashes with a previous input that's lexicographically different
from the current one.

For example:

Sanitize("_0123?foo")    = "key_0123_foo"
Sanitize("key_0123?foo") = "key_0123_foo"
Sanitize("0123_foo")	 = "key_0123_foo"

and all the above clash so it returns an error telling the caller
what and where the clash is.