Revise annotation APIs (#325)
Provide Annoate and Annoatef.

We don't want to the lazy behavior of the annotation
APIs exposed to the user. Don't document the lazy
behavior.

Passing nil as attributes meets the case for annotating
only with a message. We are removing all other functions
and methods and expect user to pass nil if they
don't want to annotate with attributes.

Fixes #129.