Add B3 propagation format support
B3 spans have two parts: client and server. For this reason,
B3 header can contain a client span ID and a parent ID of the
span at the same time.

OpenCensus trace context doesn't allow the same span ID to be
shared between client and server. This is why we are going
to assume the incoming span ID is the parent.

For similar reasons, outgoing requests will not set the parent
span ID and assume the receiver will accept the outgoing span
ID as the client side of the span.

Fixes #362.