prom: fixes dynamic registration of views bug
prometheus's default registry appears to block registration of any other
metrics based on there being >= 1 duplicate in the set of metrics returned
from Describe to register. While Describe's API seems to indicate that a
collector should return a superset of metrics, and while this makes perfect
sense with the Unregister logic, it does not seem to play nicely with us
running register on the same collector repeatedly with different metrics, nor
with calling unregister and then attempting to register everything once again.
that is to say, prometheus client seems to explicitly disallow the thing we're
seemingly trying to do. upstream: https://github.com/prometheus/client_golang/issues/47

this patch will:
* no longer call Unregister and then Register, as this is a case handled by
prometheus: https://github.com/prometheus/client_golang/blob/fcc130e101e76c5d303513d0e28f4b6d732845c7/prometheus/registry.go#L112-L117
* only return metrics that are not yet registered to prometheus registry

since ExportView appears to be called in a single thread, I am not sure
the locks are required, but have abided here. that also ruled out this
being a race of any kind.

It's also a perfectly reasonable explanation that I've not properly registered
views. fwiw am doing `v.Subscribe()` for every view in init() atm, and the
intent appears to be to be able to toggle these on/off during runtime.

I was able to generate these errors consistently on the first sweep of
metrics that would be done:

```
cannot register the collector: descriptor Desc{fqName: "redacted", help:
"redacted", constLabels: {}, variableLabels: []} already exists with the same
fully-qualified name a nd const label values
```

to be clear, this is not due to my attempting to register 2 views / metrics
under the same name, but from registerViews repeatedly registering all metrics
to prometheus registry (confirmed) which will [trivially] contain duplicates.

I apologize that I do not have a reduced reproduction. I do however have a
patch in an OSS project that will reliably reproduce this, though I hate
pointing at that as it'll take some orienting to run.

perhaps these were simply an annoyance and to be ignored, in which case we can
just remove the `onError` bits, I am seeing all metrics in the prometheus
endpoint even though I get this error spat out a lot (and, it's not clear that
this should work this way after reading the prometheus register code). there
may be other solutions :)

thanks for this library, your efforts are appreciated, am looking forward to
using it!