Revert Prometheus PR #456 but instead register only once
Fixes #469
Reverts #456

Also ensures that we only register the collector exactly
once, of which we erraneously had assumed Unregister would
do but as documented at
https://github.com/prometheus/client_golang/blob/fcc130e101e76c5d303513d0e28f4b6d732845c7/prometheus/registry.go#L89-L101
which specifies that invoking Register again even after Unregister
will just spit out an error because a collector should only
be registered once.

PR #456 attempted to fix this but didn't maintain uniqueness of
registeredViews hence doubly or more registered Prometheus metrics.
Nonetheless it was a great attempt.

With this PR, we should never get back the error that prompted
the PR in revision.