Hide Measure fields.
We don't want to expose Measurement fields or allow direct construction
of Measurement instances because we want to keep open the possibility
of adding new Measurement representations and also encourage users to
use the .M() method.

Also saw that we make some performance promises about Measure that are
not really true. Looks like there was the beginnings of an implementation
of the fast path for skipping recording but it was never fully hooked
up.