Remove NoTrace from achttp and ocgrpc integrations (#585)
NoTrace was previously used to disable trace integration. This can
now mostly be achieved by customizing the sampler that each integration
takes as an argument.

The one thing that can no longer be achieved after this PR is
disabling trace propagation while still using the plugin. I think
this is really a good thing: trace propagation should not be optional.