rpcz doesn't display any data
The change to remove view.New in favor of constructing a View
directly broke rpcz because we stopped preserving the identity
of the *View pointer.

This restores the identity of *View such that the same pointer is
always returned in view.Data (and can therefore be used in maps and
switch statements like it was in rpcz).