Fix plugin/ochttp statusCode mistype and unconditional set
Fixes #619

The mistyped source code unconditionally set variable code
to codeUnknown since
```go
var code int32
if code < 200 || code >= 400 {
   code = codeUnknown
}
```
always would set it to codeUnknown as the zero value
of int32 is 0 so always code = codeUnknown.

However if it didn't get matched in the next switch
case, e.g. a 200, we'd get an Unknown status.

Also add unit tests for status so that we never accidentally
regress.

This was discovered by demo app inspections on Stackdriver.