Proxy (http.Hijacker).Hijack() calls to Handler.responseWriter in ochttp
Fixes #642

Ensure that plugin/ochttp.Handler proxies (http.Hijacker).Hijack()
calls to its responseWriter that we captured. Also add tests to
ensure that the behavior with the Go standard library is preserved
that is:
a) HTTP/1.X connections successfully can use (http.Hijacker).Hijack()
b) HTTP/2.X connections are incompatible with (http.Hijacker).Hijack()
because with HTTP/2.X, multiple requests can be sent on the same
connection thus it doesn't make sense to hijack a connection. The
standard library also panics and enforces this behavior.

Note that both a) and b) are only enforced at runtime, like we do.
We don't need to do any work, except solidify and lock this behavior
in our tests.

As an advantage, we can now successfully use ochttp.Handler with
websockets!