Fix Prometheus Histograms to be cumulative
Fixes #649

The current Prometheus Histogram creator takes in
a mapping of bucket values to counts, having already
assumed that the creator made cumulative results i.e.
added the result of the largest minimum to the current.
Prometheus mentions this expectation at
https://godoc.org/github.com/prometheus/client_golang/prometheus#NewConstHistogram

Also added a test to ensure that we never
regress on this behavior.