unregister should properly delete View
Fixes #706
Fixes #710

Confusion that came from having two APIs:
a) Unregister
b) Unsubscribe

that initially had different purposes:
a) To completely remove the view from the underlying worker
b) To keep a reference to it but not completely remove it

However the difference between the two manifested in a bug
because we mistakenly assumed the subtle difference between
them never existed, hence on agreeing to deprecate Unsubscribe
for Unregister, the logic for Unsubscribe was kept but that
for Unregister was already lost in previous API changes in
PR https://github.com/census-instrumentation/opencensus-go/pull/492.

PR https://github.com/census-instrumentation/opencensus-go/pull/615
then naively folded Unsubscribe into Unregister and the bug then
fully manifested and could be tested by creating the same view with
a distribution aggregation in a loop and error-ing if Register fails.

With this change, I've also added a test that registers views of
different types, unregisters them in a loop and checks that we
don't error -- which is the original reproducer.
Also while here, just remove the vestiges of unsubscribeViewReq
to solidify the intent and name to unregisterViewReq.