Make plugin/ochttp/Handler implement common HTTP server interfaces
For #744

Make handler implement the following interfaces:
* http.CloseNotifier
* http.Pusher
and also adds tests to lock this behavior in and ensure
that we never regress.

In the future, we'll need to figure out why net/http is
still tripping out with http.ErrNotSupported for any
sort of server/connection combination with or without
ochttp.Handler as the handler. For now the current
code I believe satisfactorily implements the http.Pusher
interface.