tag: avoid mention of NewMap in the documentation
The NewMap function does not exist - New seems to be the
only way to create a new map.

Also lose the map size hint in newMap - it's currently always
zero and can easily be reinstated later as an optimization if needed.

Also include trivial optimization to slice allocation in String.