zpages/rpcz: fix constant error total accumulation
When there was any error on any gRPC method, the errors total kept accumulating at each interval on the /debug/rpcz page even though no additional errors were actually occurring. The Prometheus metrics looked correct, so I figured it had to be a problem with the rpcz page itself.
 
It seems a guard was put in for this already with the map, however, the map was always empty due to this bug. I verified this fix locally.