trace: add limits for attributes, message events, and links.
This applies limits to span data entities that are slices: Attributes,
Message Events, and Links.

The limits are default per the spec and can be globally changed with
trace.ApplyConfig.

The limits can also be changed per-span when creating a new span
with one or more trace.StartOption.

At the limit, it pops the oldest sample off prior to appending the new
sample.

Each test ensures 4 entities are added in total -- the first 2 are
dropped and the last 2 remain. This should prevent off by 1 bugs.