Add possibility to Jaeger to add process tags
Jaeger allows to attach tags to it's process, which this patch
allows to specify.
Currently this is limited to bool, string, int64 and int32, as
attributeToTag is used for converting it to *jaeger.Tag.

For reference see the official Jaeger client go:
https://github.com/jaegertracing/jaeger-client-go/blob/252d853b2a4f3cfc98bb89c3a8bb3d3aa50ed4d6/tracer.go#L64

This change helps to specify tags, such as 'ip', which has a special
handling in Jaeger (for example jaeger-query skips clock skew
adjustments for processes from the same IP).