zpages: fix countFormatter code and add tests
Fix countFormatter code due to a bug that
I advertently introduced with a suggestion
in PR 894, due to change of signs:
from >= to <=
it should have just been the int64 cast.

However, also fix a long standing bug in the
original code for the case num >= 12, where
the divisor was copy-pasted as 1e9 instead
of 1e12 for Tera* values.

This change also adds tests to ensure such
problems could be caught in the first place
but that we never regress.