Provde ways to avoid propagating trace and tag metadata
In certain situations it is undesirable to propagate trace or tag
metadata on outbound requests. Nevertheless, it may still be
desirable to instrument these requests using ocgrpc or ochttp so
that stats and in-process spans may be collected.

This change provides ways to accomplish this for both ocgrpc (server)
and ochttp. We still need a way to accomplish this for ocgrpc/client.