zpages: fix Snapshot int->*int64 changes + tests on count (#952)
The signature for countFormatter's arguments was changed
in PR #948 from int to int64 but unfortunately we forgot
to update the int fields from int to the new type.

Issue #948's recommendation was to use uint64 since that
affords us more time to overflow e.g. even at 25QPs
it'd take a couple of millenia to overflow with uint64.

This change completes the change of types of the int variables
but also adds tests that make use of the template functions
and a snapshot and then compare the output or fail. Previously the
tests complained but since they were HTTP tests not attached
to an output buffer that could be inspected, the bug innocently
crept away in standard output and the tests incorrectly passed as in:
* https://ci.appveyor.com/project/opencensusgoteam/opencensus-go/builds/19498802#L700
* https://travis-ci.org/census-instrumentation/opencensus-go/builds/441324783#L1183

This problem was reported by an OpenCensus-Service user in
* https://github.com/census-instrumentation/opencensus-service/issues/123

Fixes #895
Fixes #951