Add flag to force generate span data
span data will not init if span is not sampled and not enable
local span storage. so add a flag to force generate span data.
this operation will have a little more memery overhead, but
it can be use to get all information when we get a span. we can
use the information to do logging and rebuild the trace from
logs.

also change SpanData method to use makeSpanData result, because
shallow copy of span.data would allow for mutating for instance
the Attributes map (and get some concurrent access nasties).