Update LRU usage
This change updates the LRU cache implementation used in the trace
package, replacing github.com/hashicorp/golang-lru with
github.com/golang/groupcache, while preserving the manner in which trace
abstractions make use of the data structure.

The reason for this change is that github.com/hashicorp/golang-lru has a
Mozilla license, which is different from that of
github.com/census-instrumentation/opencensus-go. The difference (and the
mere presence of the former) makes it more difficult for open-source
projects (e.g., https://fuchsia.dev) to vendor the latter.